+++ cache4k/src/commonTest/kotlin/io/github/reactivecircus/cache4k/CacheLoaderTest.kt
@@ -13,7 +13,7 @@ class CacheLoaderTest {
 
     @Test
     fun entryWithAssociatedKeyNotExists_getWithLoader_returnsValueFromLoader() = runTest {
-        val cache = Cache.Builder.newBuilder()
+        val cache = Cache.Builder()
             .build<Long, String>()
 
         var loaderInvokeCount = 0
@@ -30,7 +30,7 @@ class CacheLoaderTest {
 
     @Test
     fun expiredEntryWithAssociatedKeyExists_getWithLoader_returnsValueFromLoader() = runTest {
-        val cache = Cache.Builder.newBuilder()
+        val cache = Cache.Builder()
             .expireAfterWrite(expiryDuration)
             .fakeTimeSource(fakeTimeSource)
             .build<Long, String>()
@@ -54,7 +54,7 @@ class CacheLoaderTest {
 
     @Test
     fun unexpiredEntryWithAssociatedKeyExists_getWithLoader_returnsExistingValue() = runTest {
-        val cache = Cache.Builder.newBuilder()
+        val cache = Cache.Builder()
             .expireAfterAccess(expiryDuration)
             .fakeTimeSource(fakeTimeSource)
             .build<Long, String>()
@@ -78,7 +78,7 @@ class CacheLoaderTest {
 
     @Test
     fun loaderThrowsException_getWithLoader_exceptionPropagated() = runTest {
-        val cache = Cache.Builder.newBuilder()
+        val cache = Cache.Builder()
             .build<Long, String>()
 
         var loaderInvokeCount = 0