+++ CHANGELOG.md
@@ -1,14 +1,23 @@
 # Change Log
 
+## Unreleased
+
+### Changed
+
+- Change `fun newBuilder()` to `operator fun invoke()` - [#7](https://github.com/ReactiveCircus/cache4k/pull/7).
+- Compile with Kotlin 1.5.21.
+
 ## 0.2.0
 
 ### Changed
+
 - Support Kotlin 1.5.0.
 - Stately to 1.1.7.
 
 ## 0.1.1
 
 ### Changed
+
 - Disable IR.
 
 ## 0.1.0